+++ src/OpenLoco/src/Localisation/StringIds.h
@@ -2115,6 +2115,8 @@ namespace OpenLoco::StringIds
     constexpr StringId confirm_vehicle_component_move_cargo_warning_confirm = 2401;
     constexpr StringId confirm_vehicle_component_move_cargo_warning_message = 2402;
     constexpr StringId confirm_vehicle_component_move_cargo_multiple_vehicles_warning_message = 2403;
+    constexpr StringId pause_on_lost_focus = 2404;
+    constexpr StringId pause_on_lost_focus_tip = 2405;
 
     constexpr StringId temporary_object_load_str_0 = 8192;
     constexpr StringId temporary_object_load_str_1 = 8193;