+++ src/OpenLoco/src/Ui/Windows/PromptSaveWindow.cpp
@@ -68,8 +68,6 @@ namespace OpenLoco::Ui::Windows::PromptSaveWindow
             window->flags |= Ui::WindowFlags::transparent;
 
             SceneManager::setPauseFlag(PauseFlags::promptSave);
-            Audio::pauseSound();
-            WindowManager::invalidate(WindowType::timeToolbar);
         }
 
         _savePromptType = savePromptType;
@@ -148,8 +146,6 @@ namespace OpenLoco::Ui::Windows::PromptSaveWindow
     static void onClose([[maybe_unused]] Window& self)
     {
         SceneManager::unsetPauseFlag(PauseFlags::promptSave);
-        Audio::unpauseSound();
-        WindowManager::invalidate(WindowType::timeToolbar);
     }
 
     static constexpr WindowEventList kEvents = {